Moving db initialization and config files out of src/config
