+++ src/api/app_factory.py
@@ -6,7 +6,7 @@ import re
 from src.database.db import db
 from src.config.models.user import BlacklistedToken
 
-from ..config.db_config import Config, initialize_database
+from src.database.config_db import Config, initialize_database
 
 
 def create_app():
