+++ src/config/models/user.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import UserMixin
 from src.database.db import db
-from ..db_config import Config
+from src.database.config_db import Config
 
 class User(UserMixin, db.Model):
     __tablename__ = 'user'
