+++ src/webapp/webapp.py
@@ -6,7 +6,7 @@ import logging
 from pathlib import Path
 from flask_jwt_extended import JWTManager
 from src.database.db import db
-from ..config.db_config import Config
+from src.database.config_db import Config
 from datetime import timedelta
 from src.common_utils.date_utils import parse_date
